refactor(projects): group cell files under list/cells/ subfolder
Per Ayush's round-4 review on PR #1220 (cell.tsx:1): 8 cell files flat
in list/ crowded the feature-folder top level. Moved all of them under
a cells/ subdirectory so the list/ root stays readable (just the page
shell + columns + constants + types + fake-data + the cells/ subfolder).

Layout now:
  pages/projects/list/
  ├── index.tsx           (ListView wiring)
  ├── columns.tsx
  ├── constants.ts        (PHASE_LABELS — pure data only)
  ├── types.ts
  ├── fake-data.ts
  └── cells/
      ├── index.tsx       (was cell.tsx — the dispatch)
      ├── dot.tsx
      ├── stages-icon.tsx
      ├── date-cell.tsx
      ├── project-name-cell.tsx
      ├── phase-cell.tsx
      ├── budget-progress-cell.tsx
      └── employee-cell.tsx

Used git mv for all 8 files so rename history is preserved. Import
paths updated where needed:
  - cells/index.tsx, project-name-cell.tsx, phase-cell.tsx,
    employee-cell.tsx: ./types → ../types
  - cells/phase-cell.tsx: ./constants → ../constants
  - list/index.tsx: ./cell → ./cells (resolves to cells/index.tsx)

Build clean.
