+++ frontend/packages/app/src/pages/projects/list/cells/dot.tsx
