+++ frontend/packages/app/src/pages/projects/list/cells/stages-icon.tsx
